# src/main/resources/static/openapi/mock-service.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/adeolaadebisi/VCF-CONVERTER/1.0.0
info:
  description: This is a simple API
  version: "1.0.0"
  title: Simple Inventory API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /companyinformation:
    get:
      tags:
        - VCF
      summary: searches inventory
      operationId: getCompanyInformation
      description: |
        By passing in the appropriate options, you can search for
        available inventory in the system
      parameters:
        - in: query
          name: instid
          description: pass an optional search string for looking up inventory
          required: false
          schema:
            type: string
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyInformationResponse'
        '400':
          description: bad input parameter
  /cardholderinformation:
    get:
      tags:
        - VCF
      summary: searches inventory
      operationId: getCardholderInformation
      description: |
        By passing in the appropriate options, you can search for
        available inventory in the system
      parameters:
        - in: query
          name: instid
          description: pass an optional search string for looking up inventory
          required: false
          schema:
            type: string
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardholderInformationResponse'
        '400':
          description: bad input parameter
  /accountinformation:
    get:
      tags:
        - VCF
      summary: searches inventory
      operationId: getAccountInformation
      description: |
        By passing in the appropriate options, you can search for
        available inventory in the system
      parameters:
        - in: query
          name: instid
          description: pass an optional search string for looking up inventory
          required: false
          schema:
            type: string
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInformationResponse'
        '400':
          description: bad input parameter
  /transactions:
    get:
      tags:
        - VCF
      summary: get transaction information
      operationId: getTransactionInformation
      description: |
        By passing in the appropriate options, you can search for
        available inventory in the system
      parameters:
        - in: query
          name: instid
          description: intitution id of records to fetch
          required: false
          schema:
            type: string
        - in: query
          name: startDate
          description: start date filter for record
          schema:
            type: integer
            format: int32
            minimum: 0
        - in: query
          name: endDate
          description: end date filter for record
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 50
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionInformationResponse'
        '400':
          description: bad input parameter
  /auth:
    post:
      tags:
        - VCF
      summary: adds an inventory item
      operationId: Authentication
      description: Adds an item to the system
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          description: 'invalid input, object invalid'
        '409':
          description: an existing item already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Authentication'
        description: Inventory item to add
components:
  schemas:
    AuthenticationResponse:
      type: object
      properties:
        access_token:
          type: string
          format: jwt
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJUWFVTRVJORyIsImp0aSI6ImU3ZmQzNjkzLTNlODktNDI0Mi1hNjdhLTE3NGQ3ZThjOWQyZCIsImlhdCI6IjEwLzI4LzIwMjQgMTA6MzM6MjUgQU0iLCJuYmYiOjE3MzAxMTE2MDUsImV4cCI6MTc2MTY0NzYwNSwiaXNzIjoiaHR0cHM6Ly93d3cubGlua2VkaW4uY29tL2luL3NhbXVlbC13YWluYWluYS03MWI3MDkxNzcvIiwiYXVkIjoiU2FtdWVsIFdhaW5haW5hIn0.aVDo7j8v6ckVjWKWciUlT0RhlW68YW3OjUydAitMIXw
        expires_in:
          type: integer
          example: 365
...
    CompanyInformationResponse:
      type: object
      properties:
        singleCompany:
          $ref: '#/components/schemas/CompanyInformation'
    CompanyInformation:
      type: object
      required:
        - companyID
        - companyName
        - isoCountryCode
        - fiscalYearDate
        - cardType
        - issuerName
        - organizationHierarchyManagementIndicator
      properties:
        companyID:
          type: string
          example: 41
        companyName:
          type: string
          example: 1186
        isoCountryCode:
          type: string
          example: Adeola
        fiscalYearDate:
          type: string
          example: Adebisi
        cardType:
          type: string
          example: 566
        issuerName:
          type: string
          example: 566
        organizationHierarchyManagementIndicator:
          type: string
          example: 566
    CardholderInformationResponse:
      type: object
      properties:
        cardHolderInformation:
          type: array
          items:
            $ref: '#/components/schemas/CardholderInformation'
          example:
            - companyIdentification: '41'
              cardholderIdentification: '1186'
              firstName: ALTERNATIVE1
              lastName: FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1214'
              firstName: CHARGE 1
              lastName: CHARGE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1465'
              firstName: DUNCAN
              lastName: NGARI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1219'
              firstName: CHARGE 2
              lastName: CHARGE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1364'
              firstName: AYOOLA
              lastName: AMAO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1685'
              firstName: AYOOLA
              lastName: AMAO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1384'
              firstName: MOBAYO
              lastName: ADEGBOYE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1411'
              firstName: MOBAYO
              lastName: ADEGBOYE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1471'
              firstName: BULOGOSI
              lastName: MWAVAGA
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1163'
              firstName: ALTERNATIVE
              lastName: FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1183'
              firstName: CHARGE 3
              lastName: CHARGE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1182'
              firstName: CLASSIC 1
              lastName: NORMAL 1
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1216'
              firstName: ALTERNATIVE1
              lastName: FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1218'
              firstName: CHARGE 3
              lastName: CHARGE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1322'
              firstName: CLASSIC
              lastName: ALTERNATIVE FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1344'
              firstName: AYOOLA
              lastName: AMAO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1386'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1412'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1181'
              firstName: CLASSIC 3
              lastName: NORMAL 3
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1221'
              firstName: ALTERNATIVE3
              lastName: FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1341'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1404'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1409'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1464'
              firstName: CHIKA
              lastName: ODIARI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1468'
              firstName: AYOOLA
              lastName: AMAO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1162'
              firstName: CLASSIC
              lastName: NORMAL
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1220'
              firstName: CLASSIC 2
              lastName: NORMAL 2
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1324'
              firstName: CLASSIC 1
              lastName: NORMAL 1
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1346'
              firstName: MOBAYO
              lastName: ADEGBOYE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1365'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1686'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1410'
              firstName: AYOOLA
              lastName: AMAO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1385'
              firstName: AYOOLA
              lastName: AMAO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1470'
              firstName: DENNIS
              lastName: SOITA
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1472'
              firstName: JOYCE
              lastName: NZILANI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1161'
              firstName: CLASSIC
              lastName: CHARGE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1184'
              firstName: ALTERNATIVE2
              lastName: FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1217'
              firstName: CLASSIC 1
              lastName: NORMAL 1
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1321'
              firstName: CLASSIC
              lastName: ALTERNATIVE FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1325'
              firstName: ALTERNATIVE1
              lastName: FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1188'
              firstName: CLASSIC 2
              lastName: NORMAL 2
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1222'
              firstName: CLASSIC 3
              lastName: NORMAL 3
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1323'
              firstName: CHARGE 1
              lastName: CHARGE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1345'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1467'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1514'
              firstName: CLASSIC
              lastName: TRAINING
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1645'
              firstName: DUNCAN
              lastName: NGARI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1657'
              firstName: JOYCE
              lastName: NZILANI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1665'
              firstName: JOYCE
              lastName: NZILANI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1705'
              firstName: MOBAYO
              lastName: ADEGBOYE REDO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1909'
              firstName: CLASSIC
              lastName: STANDARD
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '2248'
              firstName: DENIS
              lastName: FRED
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1651'
              firstName: CHIKA
              lastName: ODIARI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1655'
              firstName: AYOOLA
              lastName: AMAO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1670'
              firstName: DUNCAN
              lastName: NGARI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1672'
              firstName: CHIKA
              lastName: ODIARI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1704'
              firstName: AYOOLA
              lastName: AMAO REDO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1911'
              firstName: STAFF
              lastName: FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1646'
              firstName: AYOOLA
              lastName: AMAO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1653'
              firstName: DUNCAN
              lastName: NGARI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1667'
              firstName: BULOGOSI
              lastName: MWAVAGA
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1908'
              firstName: CLASSIC
              lastName: CHARGE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1944'
              firstName: STANDARD
              lastName: FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1649'
              firstName: DENNIS
              lastName: SOITA
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1644'
              firstName: BULOGOSI
              lastName: MWAVAGA
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1652'
              firstName: DENNIS
              lastName: SOITA
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1659'
              firstName: LILIAN
              lastName: OJIAMBO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1687'
              firstName: LILIAN
              lastName: OJIAMBO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1666'
              firstName: DENNIS
              lastName: SOITA
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1706'
              firstName: CHARLES 1
              lastName: NWOSU REDO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1907'
              firstName: TEST
              lastName: CARD3
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1924'
              firstName: MOBAYO
              lastName: ADEGBOYE REDO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1945'
              firstName: CLASSIC
              lastName: STAFF
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1648'
              firstName: JOYCE
              lastName: NZILANI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1658'
              firstName: CHIKA
              lastName: ODIARI
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1654'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1664'
              firstName: CHARLES 1
              lastName: NWOSU
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1671'
              firstName: LILIAN
              lastName: OJIAMBO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1926'
              firstName: AYOOLA
              lastName: AMAO REDO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1912'
              firstName: STANDARD
              lastName: FINANCE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1946'
              firstName: TEST
              lastName: CARD3
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '2052'
              firstName: DEV 2
              lastName: TESTS
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1647'
              firstName: LILIAN
              lastName: OJIAMBO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1656'
              firstName: MOBAYO
              lastName: ADEGBOYE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1669'
              firstName: AYOOLA
              lastName: AMAO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1925'
              firstName: CHARLES 1
              lastName: NWOSU REDO
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1650'
              firstName: MOBAYO
              lastName: ADEGBOYE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1688'
              firstName: BULOGOSI
              lastName: MWAVAGA
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1660'
              firstName: BULOGOSI
              lastName: MWAVAGA
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1668'
              firstName: MOBAYO
              lastName: ADEGBOYE
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '1910'
              firstName: CLASSIC
              lastName: STAFF
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '2318'
              firstName: TEST
              lastName: CARD7
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '2764'
              firstName: ALTER
              lastName: STD 3
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '4567'
              firstName: AGADA
              lastName: AMEH
              isoCountryCode: '566'
            - companyIdentification: '41'
              cardholderIdentification: '4564'
              firstName: KIZITO
              lastName: IFEANYI
              isoCountryCode: '566'
    CardholderInformation:
      type: object
      required:
        - companyIdentification
        - cardholderIdentification
        - firstName
        - lastName
        - isoCountryCode
      properties:
        companyIdentification:
          type: string
          example: 41
        cardholderIdentification:
          type: string
          example: 1186
        firstName:
          type: string
          example: Adeola
        lastName:
          type: string
          example: Adebisi
        isoCountryCode:
          type: string
          example: 566
...
    Authentication:
      type: object
      required:
        - name
        - password
      properties:
        name:
          type: string
          example: name
        password:
          type: string
          example: password
...